+++ projects/swimlane/ngx-ui/src/lib/components/json-editor/json-editor-flat/json-editor-node-flat/node-types/object-node-flat/object-node-flat.component.ts
@@ -33,7 +33,7 @@ export class ObjectNodeFlatComponent extends ObjectNode implements OnInit {
 
   @Input() schemaBuilderMode: boolean;
 
-  @Input() formats: JsonSchemaDataType[];
+  @Input() formats: JsonSchemaDataType[] = [];
 
   @Input() compressed: boolean;
 
@@ -63,7 +63,12 @@ export class ObjectNodeFlatComponent extends ObjectNode implements OnInit {
   }
 
   onUpdatePropertyName(options: { id: string; name: string }): void {
+    const oldName = this.propertyIndex[options.id].propertyName;
+    const index = Object.keys(this.schemaRef.properties).findIndex(prop => prop === oldName);
+    this.updateSchemaPropertyName(this.schemaRef, options.name, this.propertyIndex[options.id].propertyName);
+    this.swapSchemaProperties(index);
     this.updatePropertyName(options.id, options.name);
+    this.schemaChange.emit();
   }
 
   onPropertyConfig(property: JSONEditorSchema, index: number): void {
@@ -179,6 +184,7 @@ export class ObjectNodeFlatComponent extends ObjectNode implements OnInit {
     this.schemaRef.properties[prop.propertyName] = {
       type: prop.type,
       ...(prop['format'] && { format: prop['format'] }),
+      ...(prop['examples'] && { examples: prop['examples'] }),
       ...(prop['title'] && { title: prop['title'] }),
       ...(prop['items'] && { items: prop['items'] }),
       ...(prop['required'] && { required: prop['required'] }),