+++ projects/swimlane/ngx-ui/src/lib/components/overlay/index.ts
@@ -1,3 +1,4 @@
-export * from './overlay.module';
 export * from './overlay.component';
+export * from './overlay.module';
 export * from './overlay.service';
+export * from './resize-overlay.component';