+++ projects/swimlane/ngx-ui/src/lib/components/overlay/overlay.module.ts
@@ -4,12 +4,15 @@ import { CommonModule } from '@angular/common';
 import { OverlayComponent } from './overlay.component';
 import { OverlayService } from './overlay.service';
 import { InjectionService } from '../../services/injection/injection.service';
+import { ResizeOverlayComponent } from './resize-overlay.component';
+import { IconModule } from '../icon/icon.module';
+import { HotkeysService } from '../hotkeys/hotkeys.service';
 
 @NgModule({
-  declarations: [OverlayComponent],
-  providers: [OverlayService, InjectionService, OverlayService],
-  exports: [OverlayComponent],
-  imports: [CommonModule],
-  entryComponents: [OverlayComponent]
+  declarations: [OverlayComponent, ResizeOverlayComponent],
+  providers: [OverlayService, InjectionService, OverlayService, HotkeysService],
+  exports: [OverlayComponent, ResizeOverlayComponent],
+  imports: [CommonModule, IconModule],
+  entryComponents: [OverlayComponent, ResizeOverlayComponent]
 })
 export class OverlayModule {}